requirements.txt
torch
torchvision
numpy
pillow
scikit-learn
git+https://github.com/openai/CLIP.git